Moved app_ShouldNotCrash to InvalidFormTest
